requirements.txt
torch>=2.0.0
sympy>=1.11.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
tqdm>=4.64.0
python-Levenshtein>=0.20.0
scikit-learn>=1.1.0
